more errors

+++ source/commands.h
@@ -15,13 +15,13 @@ class commands {
 		/* public helper functions */
 
 		static std::string file_last_mod_time(std::string directory);
-		static std::string format_file_size(unsigned long file_size, int precision);
-		static unsigned long file_sizes(std::string directory);
-		static unsigned long file_size(std::string directory);
+		static std::string format_file_size(double file_size, int precision);
+		static double file_sizes(std::string directory);
+		static double file_size(std::string directory);
 		static int directory_items(std::string directory);
 		static std::string file_permissions(std::string directory);
 		static std::string file_owner(std::string directory);
-		static unsigned long free_space(std::string directory);
+		static double free_space(std::string directory);
 		static std::string find_and_replace(std::string str, std::string search, std::string replace);
 
 		/* main functions */
